// server/src/types.ts
import { Request, Response } from "express"
import { Session } from "express-session"
import Redis from "ioredis";
import { userLoader } from "./utils/userLoader";

export type MyContext = {
    req: Request & { session?: Session & { userId?: number } };
    res: Response;
    redis: Redis;
    userLoader: ReturnType<typeof userLoader>
}